# src/main/java/com/sc/zipdistance/model/dto/CalculateDistanceDto.java
package com.sc.zipdistance.model.dto;

import lombok.Data;

/**
 * created 18/04/2025 - 15:37
 * project zipdistance
 * author sc
 */
@Data
public class CalculateDistanceDto {

    public String postcode1;
    public String postcode2;
}