# src/main/resources/swagger/swagger-api.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/admin/users":{"put":{"tags":["user-controller"],"summary":"Update an existing user","description":"Requires 'UPDATE_USER' authority.","operationId":"updateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}},"required":true},"responses":{"403":{"description":"Forbidden","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"User updated successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseUserDto"}}}},"400":{"description":"User update failed","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseUserDto"}}}}}},"post":{"tags":["user-controller"],"summary":"Create a new user","description":"Requires 'CREATE_USER' authority.","operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}},"required":true},"responses":{"403":{"description":"Forbidden","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"User created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseUserDto"}}}},"400":{"description":"User creation failed","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseUserDto"}}}}}}},"/api/v1/admin/postcode":{"put":{"tags":["postcode-controller"],"summary":"Update a postcode","description":"Requires 'UPDATE_POSTAL' authority.","operationId":"updatePostcode","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostcodeDto"}}},"required":true},"responses":{"403":{"description":"Forbidden","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"PostCode updated successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponsePostcodeDto"}}}},"400":{"description":"PostCode update failed","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponsePostcodeDto"}}}}}},"post":{"tags":["postcode-controller"],"summary":"Create a new postcode","description":"Requires 'CREATE_POSTAL' authority.","operationId":"createPostcode","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostcodeDto"}}},"required":true},"responses":{"403":{"description":"Forbidden","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"PostCode created successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponsePostcodeDto"}}}},"400":{"description":"PostCode already exists or creation failed","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponsePostcodeDto"}}}}}}},"/api/v1/admin/postcode/calculate":{"post":{"tags":["postcode-controller"],"summary":"Calculate distance between two postcodes","description":"Requires 'CALCULATE_DISTANCE' authority.","operationId":"calculatePostcodeDistance","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CalculateDistanceDto"}}},"required":true},"responses":{"403":{"description":"Forbidden","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"Distance calculated successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseCalculateDistanceResponseDto"}}}},"400":{"description":"Unable to calculate the distance","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseCalculateDistanceResponseDto"}}}}}}},"/api/v1/admin/auth/login":{"post":{"tags":["auth-controller"],"summary":"User login","description":"Authenticates a user and returns a JWT token if successful.","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}},"required":true},"responses":{"403":{"description":"Forbidden","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"Login successful, JWT token returned","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Login failed due to invalid credentials or input","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/dev/ping":{"get":{"tags":["ping-controller"],"summary":"Health Check","operationId":"healthCheck","responses":{"403":{"description":"Forbidden","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/admin/users/{id}":{"get":{"tags":["user-controller"],"summary":"Get user by ID","description":"Requires 'VIEW_USERS' authority.","operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"403":{"description":"Forbidden","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"User retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseUserDto"}}}},"400":{"description":"User retrieval failed","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseUserDto"}}}}}}},"/api/v1/admin/postcode/{postcode}":{"get":{"tags":["postcode-controller"],"summary":"Get postcode by value","description":"Requires 'VIEW_POSTAL' authority.","operationId":"getPostcodeById","parameters":[{"name":"postcode","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"403":{"description":"Forbidden","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"PostCode retrieved successfully","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponsePostcodeDto"}}}},"404":{"description":"PostCode not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponsePostcodeDto"}}}}}}}},"components":{"schemas":{"UpdateUserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"phoneNo":{"type":"string"},"password":{"type":"string"},"confirmPassword":{"type":"string"}}},"BaseResponseUserDto":{"type":"object","properties":{"status":{"type":"string"},"message":{"type":"string"},"result":{"$ref":"#/components/schemas/UserDto"}}},"UserDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"phoneNo":{"type":"string"},"createdAt":{"type":"string"}}},"PostcodeDto":{"type":"object","properties":{"postcode":{"type":"string"},"latitude":{"type":"number"},"longitude":{"type":"number"}}},"BaseResponsePostcodeDto":{"type":"object","properties":{"status":{"type":"string"},"message":{"type":"string"},"result":{"$ref":"#/components/schemas/PostcodeDto"}}},"CreateUserDto":{"type":"object","properties":{"name":{"type":"string","minLength":1},"email":{"type":"string"},"phoneNo":{"type":"string","minLength":1},"password":{"type":"string","minLength":1},"confirmPassword":{"type":"string","minLength":1}}},"CalculateDistanceDto":{"type":"object","properties":{"postcode1":{"type":"string"},"postcode2":{"type":"string"}}},"BaseResponseCalculateDistanceResponseDto":{"type":"object","properties":{"status":{"type":"string"},"message":{"type":"string"},"result":{"$ref":"#/components/schemas/CalculateDistanceResponseDto"}}},"CalculateDistanceResponseDto":{"type":"object","properties":{"postcodeDto1":{"$ref":"#/components/schemas/PostcodeDto"},"postcodeDto2":{"$ref":"#/components/schemas/PostcodeDto"},"distance":{"type":"number","format":"double"},"unit":{"type":"string"}}},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"confirmPassword":{"type":"string"}}}}}}